backend/requirements.txt
# Phantoms Music — Backend Dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0

# YouTube audio extraction
yt-dlp>=2024.1.1
ytmusicapi>=1.3.0

# Database
aiosqlite>=0.19.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.0.0

# Utilities
pydantic>=2.0.0
email-validator>=2.0.0
Pillow>=10.0.0
